google_translator/requirements.txt
mss==9.0.2
pytesseract==0.3.8
Pillow==9.0.1
googletrans==4.0.0-rc1
PyAutoGUI==0.9.54
pynput==1.7.7
PyYAML==6.0.2
argparse==1.4.0